Removing the output stanza and moving the table writer to be a pipeline step
